# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/films/{slug}": {
            "get": {
                "description": "Get a film from a film slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Get List Example",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Film slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/lists/{user}/{slug}": {
            "get": {
                "description": "Get a list of films from a user's list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "list"
                ],
                "summary": "Get List Example",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the list owner",
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "List slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/users/{user}/watched": {
            "get": {
                "description": "Get watched fils of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get watched films per user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "letterboxd.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "is_last": {
                    "type": "boolean"
                },
                "next_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "v1.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {
                    "$ref": "#/definitions/letterboxd.Pagination"
                }
            }
        }
    }
}